requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
# faster-whisper pinned to stable version that compiles on Railway
faster-whisper==1.0.3
gTTS==2.5.1
PyMuPDF==1.24.5
requests==2.32.3
cloudinary==1.40.0
